+++ ReleaseNotes.md
@@ -26,6 +26,7 @@ SOFTWARE.
 * arm_chip_stm32: few changes using G0 family
 * arm_chip_stm32: Added other reset sources in eResetSource
 * arm_chip_stm32: pins & ports refactored to stm_xxx
+* arm_cmsis: update for STM32 families (also added G4, L5, MP1 & WB families)
 * arm_macros: SZ_ARRAY macro added
 * arm_macros: charNUL renamed to charNULL
 * arm_macros: mathematical constants under the macro form M_xxx as in math.h